+++ Tests/aweXpect.Tests/TestHelpers/Base.cs
@@ -1,0 +1,3 @@
+﻿namespace aweXpect.Tests;
+
+public class Base;
