+++ Tests/aweXpect.Tests/TestHelpers/Derived.cs
@@ -1,0 +1,6 @@
+﻿namespace aweXpect.Tests;
+
+public sealed class Derived : Base
+{
+	public string Name { get; set; } = "";
+}
