+++ Tests/aweXpect.Tests/TestHelpers/OtherDerived.cs
@@ -1,0 +1,3 @@
+﻿namespace aweXpect.Tests;
+
+public sealed class OtherDerived : Base;
