+++ Tests/aweXpect.Tests/TestHelpers/Outer.cs
@@ -1,0 +1,7 @@
+﻿namespace aweXpect.Tests;
+
+public sealed class Outer
+{
+	public Base Item { get; set; } = new Derived();
+	public Base Other { get; set; } = new Derived();
+}
